Layouts: Allow wide central column
By copying Merlot's defaults for ourselves, we can split templates for
`narrow` (default) and `wide` page layouts. This involves an amount of
fixing image-based styling they've used to accomplish this.

While we're at it, create a /posts page with links to all posts and to
the RSS feed. Icon is my own handiwork.

Replace default post with CHCon post (which admittedly needs editing) to
remove last auto-generated nonsense.

#deploy